requirements.txt
streamlit
pandas
numpy
sentence-transformers
faiss-cpu
langchain
openai
python-dotenv